Refactor: put all the POWER_SUPPLY_* knownledge into module
We don't need to expose power_supply kernel related
to uapi anymore. In this case, we will be able to use
in-VM kernels. Only remains to keep updated all the properties.
